+++ tencentcloud/iotvideo/v20211125/client.go
@@ -470,7 +470,7 @@ func NewCheckForwardAuthResponse() (response *CheckForwardAuthResponse) {
 }
 
 // CheckForwardAuth
-// 判断是否开启的转发的权限
+// 判断是否开启转发的权限
 //
 // 可能返回的错误码:
 //  INTERNALERROR = "InternalError"
@@ -481,7 +481,7 @@ func (c *Client) CheckForwardAuth(request *CheckForwardAuthRequest) (response *C
 }
 
 // CheckForwardAuth
-// 判断是否开启的转发的权限
+// 判断是否开启转发的权限
 //
 // 可能返回的错误码:
 //  INTERNALERROR = "InternalError"