+++ tencentcloud/tke/v20180525/models.go
@@ -11162,7 +11162,7 @@ type DescribeTKEEdgeClustersRequestParams struct {
 	// 最大输出条数，默认20
 	Limit *uint64 `json:"Limit,omitempty" name:"Limit"`
 
-	// 过滤条件,当前只支持按照ClusterName和云标签进行过滤
+	// 过滤条件,当前只支持按照ClusterName和云标签进行过滤,云标签过滤格式Tags:["key1:value1","key2:value2"]
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
 }
 
@@ -11179,7 +11179,7 @@ type DescribeTKEEdgeClustersRequest struct {
 	// 最大输出条数，默认20
 	Limit *uint64 `json:"Limit,omitempty" name:"Limit"`
 
-	// 过滤条件,当前只支持按照ClusterName和云标签进行过滤
+	// 过滤条件,当前只支持按照ClusterName和云标签进行过滤,云标签过滤格式Tags:["key1:value1","key2:value2"]
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
 }